modify get_op_func file

+++ 0x0F-function_pointers/3-get_op_func.c
@@ -18,7 +18,7 @@ int (*get_op_func(char *s))(int, int)
 		{"%", op_mod},
 		{NULL, NULL}
 	};
-	unsigned long int i;
+	int i;
 
 	while (ops[i].op != NULL && *(ops[i].op) != *s)
 	{
